refactor: extract ProviderPlaygroundPanel (#3501 Phase 1g)
